moved handlers into separate class
