+++ ColorPaletteGen.Bot/Program.cs
@@ -1,4 +1,5 @@
 using ColorPaletteGen.Bot;
+using ColorPaletteGen.Bot.Handlers;
 using ColorPaletteGen.Core;
 using ColorPaletteGen.Core.GenerationStrategies;
 using Telegram.Bot;
@@ -12,7 +13,8 @@ var host = Host.CreateDefaultBuilder(args)
         collection
             .AddHostedService<Worker>()
             .AddSingleton<ITelegramBotClient>(_ => new TelegramBotClient(token))
-            .AddSingleton<IUpdateHandler, UpdateHandler>()
+            .AddSingleton<UpdateHandler>()
+            .AddSingleton<IUpdateHandler, UpdateRouter>()
             .AddSingleton<IGenerationStrategy<GenerationStrategy>, RandomGenerationStrategy>()
             .AddSingleton<IGenerationStrategy<GenerationStrategy>, AnalogousGenerationStrategy>()
             .AddSingleton<ColorPaletteGenerator>();
